Reinstate set clipboard option
When performing copy, cut, and paste with commands like y, d, and p, by
default Vim uses its own location for this - the unnamed register.

Vim also offers the + and * registers to reference the system clipboard.
Note that on some systems, + and * are the same, while on others they
are different. Generally on Linux, + and  * are different: + corresponds
to the desktop clipboard (XA_SECONDARY) that is accessed using CTRL-C,
CTRL-X, and CTRL-V, while * corresponds to the X11 primary selection
(XA_PRIMARY), which stores the mouse selection and is pasted using the
middle mouse button in most applications.

With the changes in this commit, Vim synchs the * and + registers
with the unnamed register. Neovim transparently ensures that the former
two registers hold the value of the system clipboard.